pyproject.toml
[project]
name = "echo-server-bigquery"
version = "0.1.0"
description = "FastMCP Echo Server with BigQuery integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "Authlib==1.6.3",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.7",
    "cyclopts==3.24.0",
    "dnspython==2.8.0",
    "docstring_parser==0.17.0",
    "docutils==0.22",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastmcp==2.12.2",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-cloud-bigquery==3.37.0",
    "google-cloud-core==2.4.3",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "isodate==0.7.2",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "lazy-object-proxy==1.12.0",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.2",
    "mcp==1.13.1",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "packaging==25.0",
    "parse==1.20.2",
    "pathable==0.4.4",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pydantic-core==2.33.2",
    "Pygments==2.19.2",
    "pyperclip==1.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pywin32==311; sys_platform == 'win32'",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rfc3339-validator==0.1.4",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "typing-inspection==0.4.1",
    "typing-extensions==4.15.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "Werkzeug==3.1.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]